requirements.txt
streamlit==1.28.1
supabase==2.0.3
python-dotenv==1.0.0
openai==1.3.5
pillow==10.0.1
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
plotly==5.18.0
python-dateutil==2.8.2
